# src/shared/test/swagger.yml
openapi: 3.0.1
info:
  title: U-Kitchen API
  description: 'API del sistema de restaurante.'
  servers:
    - url: http://localhost:3000/api-docs
...
paths:
...
  /supplier/findAll:
    get:
      tags:
        - supplier
      summary: Get all suppliers
      operationId: findAllSupplier
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'
        '400':
          description: Invalid status value
          content: {}
...
  /supplier/{id}:
    put:
      tags:
        - supplier
      summary: Updates a supplier in the db with form data
      operationId: updateSupplier
      parameters:
        - name: id
          in: path
          description: ID of supplier that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/SupplierIdDto'
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UpdateSupplierDto'
      responses:
        '405':
          description: Invalid input
          content: {}
    delete:
      tags:
        - supplier
      summary: Deletes a supplier
      operationId: removeSupplier
      parameters:
        - name: id
          in: path
          description: Supplier id to delete
          required: true
          schema:
            $ref: '#/components/schemas/SupplierIdDto'
      responses:
        '404':
          description: Supplier not found
          content: {}
...
  /table/findAll:
    get:
      tags:
        - table
      summary: Get all tables
      operationId: findAllTable
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Table'
        '400':
          description: Invalid status value
          content: {}
...
  /table/{id}:
    put:
      tags:
        - table
      summary: Updates a table in the db with form data
      operationId: updateTable
      parameters:
        - name: id
          in: path
          description: ID of table that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/TableIdDto'
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UpdateTableDto'
      responses:
        '405':
          description: Invalid input
          content: {}
    delete:
      tags:
        - table
      summary: Deletes a table
      operationId: removeTable
      parameters:
        - name: id
          in: path
          description: Table id to delete
          required: true
          schema:
            $ref: '#/components/schemas/TableIdDto'
      responses:
        '404':
          description: Table not found
          content: {}
...
  /order/findAllClientOrders/{id}:
    get:
      tags:
        - order
      summary: Finds all orders by client ID
      description: Returns a list of all orders for a given client ID
      operationId: findAllOrdersByClientId
      parameters:
        - name: id
          in: path
          description: ID of client to find orders for
          required: true
          schema:
            $ref: '#/components/schemas/ClientIdDto'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status value
          content: {}
        '404':
          description: Client not found
          content: {}
...
  /user/{id}:
    put:
      tags:
        - user
      summary: Updates an user in the db with form data
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID of user that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UserIdDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
        required: true
      responses:
        '405':
          description: Invalid input
          content: {}
    delete:
      tags:
        - user
      summary: Deletes an user
      operationId: removeUser
      parameters:
        - name: id
          in: path
          description: User id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UserIdDto'
      responses:
        '404':
          description: User not found
          content: {}
components:
  schemas:
...
    ClientIdDto:
      type: string
      required:
        - id
...
    Order:
      type: object
      properties:
        description:
          type: string
          example: 'Pedido para mesa 12 con preferencia sin sal'
        status:
          type: string
          example: 'pending'
        estimatedEndTime:
          type: string
          format: date-time
          example: '2025-08-08T15:00:00Z'
        endTime:
          type: string
          format: date-time
          example: '2025-08-08T15:10:00Z'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        client:
          type: object
          properties:
            id:
              type: string
              example: '1954371590753882225'
          description: ID del cliente que realizó el pedido
        table:
          type: object
          properties:
            id:
              type: string
              example: '1954371590753882225'
          description: ID de la mesa del pedido
...
    OrderItem:
      type: object
      properties:
        dish:
          type: object
          properties:
            id:
              type: string
              example: '1954707475630395502'
        quantity:
          type: integer
          example: 2
...
    Supplier:
      type: object
      properties:
        companyName:
          type: string
          example: 'La fabrica de humo'
        taxId:
          type: string
          example: '30-12345678-9'
        mail:
          type: string
          format: email
          example: 'contacto@ladrilloenlasmanos.com'
        phoneNumber:
          type: string
          example: '+54 9 341 555 1234'
        typeIngredient:
          type: string
          example: 'Pasta'
        fullName:
          type: string
          example: 'Geronimo Benavides'
        bussinessName:
          type: string
          example: 'Politica Evitemos S.A.'
    SupplierIdDto:
      type: string
      required:
        - id:
    UpdateSupplierDto:
      allOf:
        - $ref: '#/components/schemas/Supplier'
      description: 'Todos los campos son opcionales para actualización'
    Table:
      type: object
      properties:
        cod:
          type: string
        capacity:
          type: integer
        description:
          type: string
        occupied:
          type: boolean
          example: false
        sector:
          type: string
          example: sector 1
      required:
        - cod
        - capacity
        - occupied
    TableIdDto:
      type: string
      required:
        - id
    UpdateTableDto:
      allOf:
        - $ref: '#/components/schemas/Table'
      description: DTO para actualizar una mesa, todos los campos son opcionales
    User:
      type: object
      properties:
        email:
          type: string
          example: 'admin@admin.com'
        fullName:
          type: string
          example: 'Valentin'
        password:
          type: string
          format: password
          example: 'ValenPro123'
        phoneNumber:
          type: string
          example: '+54 9 341 254 7935'
        role:
          type: string
          example: 'admin'
        profilePicture:
          type: string
        client:
          type: object
          properties:
            id:
              type: string
              example: '1954371590753882225'
        employee:
          type: object
          properties:
            id:
              type: string
              example: '1954371590753882225'
      required:
        - email
        - fullName
        - password
        - phoneNumber
        - role            
    UserIdDto:
      type: string
      required:
        - id
    UpdateUserDto:
      allOf:
        - $ref: '#/components/schemas/User'

  # securitySchemes:
  #   petstore_auth:
  #     type: oauth2
  #     flows:
  #       implicit:
  #         authorizationUrl: http://petstore.swagger.io/oauth/dialog
  #         scopes:
  #           write:pets: modify pets in your account
  #           read:pets: read your pets
  #   api_key:
  #     type: apiKey
  #     name: api_key
  #     in: header
...